Update libc_errno to work correctly for both overlay and full build modes.

+++ libc/test/src/time/gmtime_test.cpp
@@ -26,7 +26,7 @@ TEST(LlvmLibcGmTime, OutOfRange) {
                         TimeConstants::NUMBER_OF_SECONDS_IN_LEAP_YEAR);
   struct tm *tm_data = LIBC_NAMESPACE::gmtime(&seconds);
   EXPECT_TRUE(tm_data == nullptr);
-  EXPECT_EQ(libc_errno, EOVERFLOW);
+  EXPECT_EQ(static_cast<int>(libc_errno), EOVERFLOW);
 
   libc_errno = 0;
   seconds = INT_MIN * static_cast<int64_t>(
@@ -34,7 +34,7 @@ TEST(LlvmLibcGmTime, OutOfRange) {
             1;
   tm_data = LIBC_NAMESPACE::gmtime(&seconds);
   EXPECT_TRUE(tm_data == nullptr);
-  EXPECT_EQ(libc_errno, EOVERFLOW);
+  EXPECT_EQ(static_cast<int>(libc_errno), EOVERFLOW);
 }
 
 TEST(LlvmLibcGmTime, InvalidSeconds) {
